pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "karlia"
description = 'Karlia python client'
authors = [
    {name = "Atelier Blanche", email = "contact@atelier-blanche.paris"},
]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.10"

dynamic = ["version"]

dependencies = [
  'httpx',
  'pydantic',
]

[project.optional-dependencies]
dev = [
  # unittests
  'coverage>=5.0.3',
  'pytest',
  'pytest-asyncio',
  'vcrpy',

  # qa
  'flake8',
  'flake8-bugbear',
  'flake8-builtins',
  'black',
  'isort',
  'pre-commit',
]

[tool.isort]
line_length = 88
profile = 'black'

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | build
  )/
)
'''

[tool.mypy]
python_version = 3.10

ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_configs = true
show_column_numbers = true

disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false

no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = ['tests.*']
disallow_subclassing_any = false
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false

[tool.pytest.ini_options]
asyncio_mode = 'auto'

[tool.hatch.version]
path = "karlia/__init__.py"
